siihdmi: inline siihdmi_detect_sink
The framebuffer routines also work with the cea861 data.  Move the edid reading
into the function itself, which avoids us having to read the EDID data multiple
times and allows us to pass the full EDID data to the framebuffer routines.

We always have to read the first block to determine the actual length of the
EDID data.  In order to keep the reading logic the same all the time, we simply
re-read the EDID data always (even if there are no extension blocks).  This
hopefully will help keep timing issues similar across all devices.