siihdmi: better IPU settling handling
For some reason pausing to allow the IPU to catchup allows the driver to
interact with a wider variety of devices.  It is still unknown as to the
concrete reason why it improves compatibility.  Unify the definition for
the timeout and only do it on the Efika MX for the time being.