siihdmi: add siihdmi_platform_data
SII9022 requires some platform specific data that was previously hardcoded into
the driver.  Create a new siihdmi_platform_data struct to actually get this
information.  This allows a particular vendor to supply the IRQ vector, the
reset method, the framebuffer id, as well as the vendor/description for the SPD
block currently.