ENGR00138516: MX5X: Initialize CPU working point correctly for DVFS-CORE
CPU working point needs to be initialized based on the current CPU
frequency in start_dvfs() and should correspond with the correct entry
in the cpu_wp_auto table.

If its not initialized correctly, DVFS-CORE will fail when PLL-relock
results in a frequency that is not the same as set in the cpu_wp_auto table.
So fix it by finding the entry that closely matches the CPU frequency.

Signed-off-by: Ranjani Vaidyanathan <ra5478@freescale.com>