sbs: dont leak memory
The power supply system does not actually expect allocated strings that it will
manage.  We were allocating the strings, expecting the power supply class driver
to clean them up, and as a result had a massive leak.  Fix this by storing the
string in the cache and returning a reference to that string.

This effectively overhauls the sbs_battery_field.  It is renamed
sbs_battery_register which is more accurate.  There is now string handling logic
for the string reading case.  The int/uint types are handled the same, so there
is a single read type for those two value types.