sbs: get battery info on first request
It seems that performing the battery info query too early can potentially cause
a race with the i2c layer.  batt->platform ends up being NULL for some reason,
and a kernel oops ensues.  Instead, just track the state of the info fields, and
if it is not valid on the first data access, just fetch the data then.