amd-gpu: reintroduce fine-grained locking
This was removed in the initial stages of moving to a less abstracted
driver since Freescale never enabled it at the time. However, since
that time it's been enabled to improve stability, so we're putting it
back now to match Freescale.

Some of the mutex initialization I think is being done in a slightly
non-standard way (but the same way it's done in the Freescale kernel
since it's taken right from the old abstraction and just opencoded
instead of being a set of defines). Will investigate this later.