amd-gpu: use Qualcomm's register definitions
This is a BIG hack.

Logging (as in kgsl_log.c) won't work anymore as it uses some weird
way to define registers based on the old register definitions.

Some FBC interrupt thing is really odd, there are no docs for it..

MMU config may be a little screwy - or not :)

Freescale's kernel driver does some things the Qualcomm driver
doesn't in some registers Qualcomm doesn't use. Odd, but kept the
FSL way since userspace may need to match and it's not worth the
work right now.