amd-gpu: wean g12 and yamato parts off kgsl_hwaccess
Also while we're at it don't bother using the device table and many
levels of dereference to just call regread/regwrite.

No functional difference, although tiger now is not as happy as it
once was when running on imxng - on fbdev it's absolutely fine (and
marginally faster over 10 runs).

At some point we're going to find this damned race condition :]

As we are poking around, modify the Yamato idle check to use a
timeout just like QCOM do. Also, no functional difference (never
seems to time out)