amd-gpu: go for a global include so userspace can eventually compile against it
While we're doing that rationalize some namings. This actually gives us a good
indication of what is used by userspace and what is not - although a few things
like flags in kgsl_types.h may still be referenced, for now we don't move them
(since Qualcomm's driver doesn't expose them to userspace, most of them are
odd power control and/or memdesc definition).

We lose two includes from the main tree (ioctl and properties) since they are
global and for userspace.