amd-gpu: remove unused-by-userspace properties
KGSL_PROP_SHMEM and KGSL_PROP_SHMEM_APERTURES are not used in the FSL
userspace. So just get rid of them. DMI properties got removed a while
back so remove it's definition from mxc_kgsl.h.

The only two properties it cares about are shadow memory and the
device info, which matches Qualcomm's driver. CAVEAT FOLLOWS.

There is a weird usage of the power property INTERNALLY which is an
odd way to do it and will be coded out eventually. Userspace does
not touch these properties, only the odd internal kernel usage..